sysdeps: add sys_stat() and sys_clock_get() on linux, move stat.h to abi-bits
