+++ sysdeps/linux/include/abi-bits/stat.h
@@ -1,0 +1,1 @@
+../../../../abis/linux/x86_64/stat.h
