+++ sysdeps/linux/meson.build
@@ -14,6 +14,7 @@ if not no_headers
 		'include/abi-bits/seek-whence.h',
 		'include/abi-bits/vm-flags.h',
 		'include/abi-bits/errno.h',
+		'include/abi-bits/stat.h',
 		subdir: 'abi-bits'
 	)
 endif
