+++ sysdeps/qword/meson.build
@@ -17,6 +17,7 @@ if not no_headers
 		'include/abi-bits/seek-whence.h',
 		'include/abi-bits/vm-flags.h',
 		'include/abi-bits/errno.h',
+		'include/abi-bits/stat.h',
 		subdir: 'abi-bits'
 	)
 	install_headers(
