+++ sysdeps/sigma/include/abi-bits/stat.h
@@ -1,0 +1,1 @@
+../../../../abis/mlibc/stat.h
